requirement.txt
Django==3.2.4
django-environ==0.4.5
psycopg2-binary==2.8.5
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0